fix bug migration

+++ database/migrations/2025_11_19_142544_create_posts_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
-            $table->integer('post_id');
-            $table->integer('owner_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('owner_id');
             $table->string('text', 2000);
             $table->string('link');
             $table->json('attachments');
